Добавлено первое заклинание
Добавлено заклинание сильный удар и пофикшены хитбоксы. Отображение хитбоксов убрано пока что((((. Системы маны еще нет.

+++ NTO_Soulslike/Assets/Scripts/Enemy.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class Enemy : MonoBehaviour
 {
-    public int HP;
-    public void TakeDamage(int damage)
+    public float HP;
+    public void TakeDamage(float damage)
     {
         HP -= damage;
         Debug.Log(HP);
